Move expand_paths to IO utils
